refactor: api폴더명 apis로 통일
