+++ apps/client/src/pages/all-memo/all-memo-page.tsx
@@ -3,7 +3,7 @@ import {
   type MemoListViewHelpers,
 } from '@shared/features/memo-list-view';
 
-import { useGetAllMemo, useGetMemoTotalCount } from './api/queries';
+import { useGetAllMemo, useGetMemoTotalCount } from './apis/queries';
 import AllMemoEmptyView from './components/all-memo-empty-view/all-memo-empty-view';
 
 const AllMemoPage = () => {
