+++ apps/client/src/pages/all-memo/apis/end-point.ts
