+++ apps/client/src/pages/all-memo/apis/queries.ts
