+++ apps/client/src/shared/features/ai-prompt/components/ai-message-bubble/ai-message-bubble.tsx
