+++ apps/client/src/shared/features/ai-prompt/components/ai-message/ai-message.css.ts
