+++ apps/client/src/shared/features/ai-prompt/components/ai-message/ai-message.tsx
