--- apps/client/src/shared/features/ai-prompt/components/index.ts
@@ -1,16 +1,0 @@
-export type {
-  AiOption,
-  Message,
-  MessageType,
-  SelectedMemo,
-  UseAiPromptProps,
-} from './model/types';
-export { useAiPrompt } from './model/use-ai-prompt';
-export { default as AiMessage } from './ui/ai-message/ai-message';
-export { default as AiMessageBubble } from './ui/ai-message-bubble/ai-message-bubble';
-export { default as PromptHeader } from './ui/prompt-header/prompt-header';
-export { default as PromptInput } from './ui/prompt-input/prompt-input';
-export { default as PromptMessagesList } from './ui/prompt-messages-list/prompt-messages-list';
-export { default as PromptOption } from './ui/prompt-option/prompt-option';
-export { default as SelectedMemosList } from './ui/selected-memo-section/selected-memo-section';
-export { default as UserMessageBubble } from './ui/user-message-bubble/user-message-bubble';
