+++ apps/client/src/shared/features/ai-prompt/components/prompt-header/prompt-header.tsx
