+++ apps/client/src/shared/features/ai-prompt/components/prompt-input/prompt-input.tsx
