+++ apps/client/src/shared/features/ai-prompt/components/prompt-option/prompt-option-item.tsx
