+++ apps/client/src/shared/features/ai-prompt/components/prompt-option/prompt-option.css.ts
