+++ apps/client/src/shared/features/ai-prompt/components/prompt-option/prompt-popover.tsx
