+++ apps/client/src/shared/features/ai-prompt/components/selected-memo-section/selected-memo-section.tsx
