+++ apps/client/src/shared/features/ai-prompt/components/user-message-bubble/user-message-bubble.tsx
