+++ apps/client/src/shared/features/ai-prompt/model/types.ts
