+++ apps/client/src/shared/features/memo-list-view/apis/end-point.ts
