+++ apps/client/src/shared/features/memo-list-view/apis/queries.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { ALL_MEMO_KEY } from '@pages/all-memo/api/query-key';
+import { ALL_MEMO_KEY } from '@pages/all-memo/apis/query-key';
 
 import { api } from '@shared/apis/instance';
