+++ apps/client/src/shared/features/memo-list-view/components/memo-list/memo-card-grid.tsx
@@ -3,10 +3,10 @@ import { useQueryClient } from '@tanstack/react-query';
 
 import { Card, DetailModal } from '@cds/ui';
 
-import { ALL_MEMO_KEY } from '@pages/all-memo/api/query-key';
+import { ALL_MEMO_KEY } from '@pages/all-memo/apis/query-key';
 
 import type { MockMemo } from '@shared/features/memo-list-view/components/memo-list/types/memo';
-import { useDetailMemo } from '@shared/features/memo-list-view/components/tree-view/components/tree-memo/api/queries';
+import { useDetailMemo } from '@shared/features/memo-list-view/components/tree-view/components/tree-memo/apis/queries';
 import useSingleAndDoubleClick from '@shared/hooks/use-single-and-double-click';
 
 import * as styles from './memo-card-grid.css';
