+++ apps/client/src/shared/features/memo-list-view/components/tree-view/components/tree-memo/tree-memo.tsx
@@ -7,7 +7,7 @@ import useSingleAndDoubleClick from '@shared/hooks/use-single-and-double-click';
 import { LabelTextType } from '@shared/types/label-type';
 import { StructureMemoTypes } from '@shared/types/memo-info-type';
 
-import { useDetailMemo } from './api/queries';
+import { useDetailMemo } from './apis/queries';
 
 import * as styles from './tree-memo.css';
